Add separate header for each component
