+++ atcg_lib/include/Scene/Components.h
@@ -1,594 +1,22 @@
 #pragma once
 
-
-#include <Core/glm.h>
-#include <Core/UUID.h>
-#include <Renderer/Shader.h>
-#include <Renderer/Camera.h>
-#include <Renderer/Renderer.h>
-#include <Renderer/Material.h>
-#include <DataStructure/Graph.h>
-#include <nanort.h>
-#include <Scripting/Script.h>
-#include <Asset/AssetManagerSystem.h>
-#include <DataStructure/BoundingBox.h>
-#include <Scripting/Behavior.h>
-
-#include <vector>
-
-namespace atcg
-{
-struct TransformComponent
-{
-    /**
-     * @brief Create a transform component
-     *
-     * @param position The position
-     * @param scale The scale
-     * @param rotation Rotation in Euler angles
-     */
-    TransformComponent(const glm::vec3& position = glm::vec3(0),
-                       const glm::vec3& scale    = glm::vec3(1),
-                       const glm::vec3& rotation = glm::vec3(0))
-        : _position(position),
-          _scale(scale),
-          _rotation(rotation)
-    {
-        calculateModelMatrix();
-    }
-
-    /**
-     * @brief Create a transform component
-     *
-     * @param model The model matrix
-     */
-    TransformComponent(const glm::mat4& model) : _model_matrix(model) { decomposeModelMatrix(); }
-
-    /**
-     * @brief Set the position
-     *
-     * @param position The position
-     */
-    ATCG_INLINE void setPosition(const glm::vec3& position)
-    {
-        _position = position;
-        ++_revision;
-        calculateModelMatrix();
-    }
-
-    /**
-     * @brief Set the rotation (Euler angles)
-     *
-     * @param rotation The euler angles
-     */
-    ATCG_INLINE void setRotation(const glm::vec3& rotation)
-    {
-        _rotation = rotation;
-        ++_revision;
-        calculateModelMatrix();
-    }
-
-    /**
-     * @brief Set the scale
-     *
-     * @param scale The scale
-     */
-    ATCG_INLINE void setScale(const glm::vec3& scale)
-    {
-        _scale = scale;
-        ++_revision;
-        calculateModelMatrix();
-    }
-
-    /**
-     * @brief Set the model matrix
-     *
-     * @param model The model matrix
-     */
-    ATCG_INLINE void setModel(const glm::mat4& model)
-    {
-        _model_matrix = model;
-        ++_revision;
-        decomposeModelMatrix();
-    }
-
-    /**
-     * @brief Get the model matrix
-     * @return The model matrix
-     */
-    ATCG_INLINE glm::mat4 getModel() const { return _model_matrix; }
-
-    /**
-     * @brief Get the position
-     * @return The position
-     */
-    ATCG_INLINE glm::vec3 getPosition() const { return _position; }
-
-    /**
-     * @brief Get the scale
-     * @return The scale
-     */
-    ATCG_INLINE glm::vec3 getScale() const { return _scale; }
-
-    /**
-     * @brief Get the rotation
-     * @return The rotation
-     */
-    ATCG_INLINE glm::vec3 getRotation() const { return _rotation; }
-
-    /**
-     * @brief Get the number of revision (edits) of the transform component
-     * @return The number of revisions
-     */
-    ATCG_INLINE uint64_t revision() const { return _revision; }
-
-    ATCG_INLINE operator glm::mat4() const { return _model_matrix; }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Transform"; }
-
-private:
-    void calculateModelMatrix();
-    void decomposeModelMatrix();
-
-private:
-    glm::mat4 _model_matrix = glm::mat4(1);
-    glm::vec3 _position     = glm::vec3(0);
-    glm::vec3 _scale        = glm::vec3(1);
-    glm::vec3 _rotation     = glm::vec3(0);    // Euler Angles
-    uint64_t _revision      = 0;
-};
-
-struct IDComponent
-{
-    IDComponent() : _ID(UUID()) {}
-    IDComponent(uint64_t id) : _ID(UUID(id)) {}
-
-    ATCG_INLINE UUID ID() const { return _ID; }
-
-private:
-    UUID _ID;
-};
-
-struct NameComponent
-{
-    NameComponent() = default;
-    NameComponent(const std::string& name) : _name(name) {}
-
-    ATCG_INLINE const std::string& name() const { return _name; }
-
-private:
-    std::string _name;
-};
-
-struct GeometryComponent
-{
-    GeometryComponent() = default;
-    GeometryComponent(AssetHandle handle) : graph_handle(handle) {}
-    GeometryComponent(const atcg::ref_ptr<Graph>& graph)
-    {
-        if(AssetManager::isAssetHandleValid(graph->handle))
-        {
-            graph_handle = graph->handle;
-        }
-        else
-        {
-            graph_handle = AssetManager::registerAsset(graph, "graph");
-        }
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Graph> graph() const { return AssetManager::getAsset<Graph>(graph_handle); }
-
-    AssetHandle graph_handle;
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Geometry"; }
-};
-
-struct AccelerationStructureComponent
-{
-    AccelerationStructureComponent() = default;
-
-    // Don't retrieve this from opengl each time used
-    atcg::MemoryBuffer<glm::vec3> vertices;
-    atcg::MemoryBuffer<glm::u32vec3> faces;
-    nanort::BVHAccel<float> accel;
-};
-
-struct CameraComponent
-{
-    CameraComponent() = default;
-    CameraComponent(const atcg::ref_ptr<Camera>& camera, const uint32_t width = 1024, const uint32_t height = 1024)
-        : camera(camera),
-          width(width),
-          height(height)
-    {
-        if(dynamic_cast<PerspectiveCamera*>(camera.get()))
-        {
-            perspective = true;
-        }
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Texture2D> image() const
-    {
-        auto image = AssetManager::getAsset<Texture2D>(image_handle);
-        return image;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Camera"; }
-
-    atcg::ref_ptr<Camera> camera;
-    glm::vec3 color    = glm::vec3(1);
-    bool perspective   = false;
-    uint32_t width     = 1024;
-    uint32_t height    = 1024;
-    float render_scale = 1.0f;
-
-    atcg::ref_ptr<atcg::Framebuffer> preview;
-    AssetHandle image_handle;
-    bool render_preview = false;
-};
-
-struct EditorCameraComponent : public CameraComponent
-{
-    EditorCameraComponent() = default;
-    EditorCameraComponent(const atcg::ref_ptr<Camera>& camera) : CameraComponent(camera) {}
-};
-
-struct RenderComponent
-{
-    RenderComponent() { default_material = atcg::make_ref<OpaqueMaterial>(); }
-
-    bool visible = true;
-
-    atcg::ref_ptr<Material> default_material;
-};
-
-struct MeshRenderComponent : public RenderComponent
-{
-    MeshRenderComponent() : RenderComponent() {}
-    MeshRenderComponent(const atcg::ref_ptr<Shader>& shader) : RenderComponent()
-    {
-        if(AssetManager::isAssetHandleValid(shader->handle))
-        {
-            shader_handle = shader->handle;
-        }
-        else
-        {
-            shader_handle = AssetManager::registerAsset(shader, "shader");
-        }
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Material> material() const
-    {
-        auto mat = AssetManager::getAsset<Material>(material_handle);
-        return mat ? mat : default_material;
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Shader> shader() const
-    {
-        auto shader = AssetManager::getAsset<Shader>(shader_handle);
-        return shader ? shader : default_shader;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Mesh Renderer"; }
-
-    atcg::ref_ptr<Shader> default_shader = atcg::ShaderManager::getShader("base");
-    bool receive_shadow                  = true;
-
-    AssetHandle material_handle = 0;
-    AssetHandle shader_handle   = 0;
-};
-
-struct PointRenderComponent : public RenderComponent
-{
-    PointRenderComponent(const glm::vec3& color = glm::vec3(1), const float& point_size = 1.0f)
-        : RenderComponent(),
-          color(color),
-          point_size(point_size)
-    {
-    }
-    PointRenderComponent(const atcg::ref_ptr<Shader>& shader,
-                         const glm::vec3& color  = glm::vec3(1),
-                         const float& point_size = 1.0f)
-        : RenderComponent(),
-          color(color),
-          point_size(point_size)
-    {
-        if(AssetManager::isAssetHandleValid(shader->handle))
-        {
-            shader_handle = shader->handle;
-        }
-        else
-        {
-            shader_handle = AssetManager::registerAsset(shader, "shader");
-        }
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Shader> shader() const
-    {
-        auto shader = AssetManager::getAsset<Shader>(shader_handle);
-        return shader ? shader : default_shader;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Point Renderer"; }
-
-    atcg::ref_ptr<Shader> default_shader = atcg::ShaderManager::getShader("base");
-    glm::vec3 color                      = glm::vec3(1);
-    float point_size                     = 1.0f;
-
-    AssetHandle shader_handle = 0;
-};
-
-struct PointSphereRenderComponent : public RenderComponent
-{
-    PointSphereRenderComponent(const float& point_size = 0.1f) : RenderComponent(), point_size(point_size) {}
-
-    PointSphereRenderComponent(const atcg::ref_ptr<Shader>& shader, const float& point_size = 0.1f)
-        : RenderComponent(),
-          point_size(point_size)
-    {
-        if(AssetManager::isAssetHandleValid(shader->handle))
-        {
-            shader_handle = shader->handle;
-        }
-        else
-        {
-            shader_handle = AssetManager::registerAsset(shader, "shader");
-        }
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Material> material() const
-    {
-        auto mat = AssetManager::getAsset<Material>(material_handle);
-        return mat ? mat : default_material;
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Shader> shader() const
-    {
-        auto shader = AssetManager::getAsset<Shader>(shader_handle);
-        return shader ? shader : default_shader;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Point Sphere Renderer"; }
-
-    atcg::ref_ptr<Shader> default_shader = atcg::ShaderManager::getShader("base");
-    float point_size                     = 0.1f;
-
-    AssetHandle material_handle = 0;
-    AssetHandle shader_handle   = 0;
-};
-
-struct EdgeRenderComponent : public RenderComponent
-{
-    EdgeRenderComponent(const glm::vec3& color = glm::vec3(1)) : RenderComponent(), color(color) {}
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Edge Renderer"; }
-
-    glm::vec3 color = glm::vec3(1);
-    // TODO Edge radius?
-};
-
-struct EdgeCylinderRenderComponent : public RenderComponent
-{
-    EdgeCylinderRenderComponent(float radius = 0.001f) : RenderComponent(), radius(radius) {}
-
-    ATCG_INLINE atcg::ref_ptr<Material> material() const
-    {
-        auto mat = AssetManager::getAsset<Material>(material_handle);
-        return mat ? mat : default_material;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Edge Cylinder Renderer"; }
-
-    float radius = 0.001f;
-
-    AssetHandle material_handle = 0;
-};
-
-struct InstanceRenderComponent : public RenderComponent
-{
-    InstanceRenderComponent() : RenderComponent() {}
-
-    void addInstanceBuffer(const atcg::ref_ptr<VertexBuffer>& instance_vbo) { instance_vbos.push_back(instance_vbo); }
-
-    ATCG_INLINE atcg::ref_ptr<Material> material() const
-    {
-        auto mat = AssetManager::getAsset<Material>(material_handle);
-        return mat ? mat : default_material;
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Shader> shader() const
-    {
-        auto shader = AssetManager::getAsset<Shader>(shader_handle);
-        return shader ? shader : default_shader;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Instance Renderer"; }
-
-    std::vector<atcg::ref_ptr<VertexBuffer>> instance_vbos;
-    atcg::ref_ptr<atcg::Shader> default_shader = nullptr;
-    bool receive_shadow                        = true;
-
-    AssetHandle material_handle = 0;
-    AssetHandle shader_handle;
-};
-
-struct CustomRenderComponent : public RenderComponent
-{
-    using RenderCallbackFn = std::function<void(Entity, const atcg::ref_ptr<Camera>& camera)>;
-
-    CustomRenderComponent(const RenderCallbackFn& callback) : RenderComponent(), callback(callback) {}
-
-    RenderCallbackFn callback;
-};
-
-struct PointLightComponent
-{
-    PointLightComponent(const float intensity = 1.0f, const glm::vec3& color = glm::vec3(1))
-        : intensity(intensity),
-          color(color)
-    {
-    }
-
-    float intensity  = 1.0f;
-    glm::vec3 color  = glm::vec3(1);
-    bool cast_shadow = true;
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Point Light"; }
-};
-
-struct MeshLightComponent
-{
-    MeshLightComponent()
-    {
-        glm::vec3 color(1);
-        TextureSpecification spec_emissive;
-        spec_emissive.width  = 1;
-        spec_emissive.height = 1;
-        glm::u8vec4 color_quant((uint8_t)(color[0] * 255.0f),
-                                (uint8_t)(color[1] * 255.0f),
-                                (uint8_t)(color[2] * 255.0f),
-                                (uint8_t)(255.0f));
-        _emissive_texture = atcg::Texture2D::create(&color_quant, spec_emissive);
-    }
-
-    float intensity = 1.0f;
-
-    ATCG_INLINE void setEmissiveColor(const glm::vec3& color)
-    {
-        TextureSpecification spec_emissive;
-        spec_emissive.width  = 1;
-        spec_emissive.height = 1;
-        glm::u8vec4 color_quant((uint8_t)(color[0] * 255.0f),
-                                (uint8_t)(color[1] * 255.0f),
-                                (uint8_t)(color[2] * 255.0f),
-                                (uint8_t)(255.0f));
-        _emissive_texture = atcg::Texture2D::create(&color_quant, spec_emissive);
-    }
-
-    ATCG_INLINE atcg::ref_ptr<atcg::Texture2D> getEmissiveTexture() const
-    {
-        auto texture = AssetManager::getAsset<Texture2D>(emissive_handle);
-        return texture ? texture : _emissive_texture;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Mesh Light"; }
-
-    AssetHandle emissive_handle;
-
-private:
-    atcg::ref_ptr<atcg::Texture2D> _emissive_texture;
-};
-
-struct ScriptComponent
-{
-    ScriptComponent() = default;
-
-    ScriptComponent(const atcg::ref_ptr<Script>& script) { setScript(script); }
-
-    ATCG_INLINE atcg::ref_ptr<Script> script() const { return AssetManager::getAsset<Script>(script_handle); }
-
-    ATCG_INLINE void setScript(const atcg::ref_ptr<Script>& script)
-    {
-        if(AssetManager::isAssetHandleValid(script->handle))
-        {
-            script_handle = script->handle;
-        }
-        else
-        {
-            script_handle = AssetManager::registerAsset(script, "script");
-        }
-        _behavior = nullptr;
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Behavior>
-    behavior(const atcg::ref_ptr<Scene>& scene, atcg::Entity entity, bool recreate = false)
-    {
-        auto scr = script();
-
-        if(!scr) return nullptr;
-
-        if(recreate || !_behavior)
-        {
-            ATCG_DEBUG("Create new behavior");
-            _behavior = scr->createBehavior(scene, entity);
-        }
-
-        return _behavior;
-    }
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Script"; }
-
-    AssetHandle script_handle = 0;
-
-private:
-    atcg::ref_ptr<Behavior> _behavior = nullptr;
-};
-
-struct HomogeneousMediumComponent
-{
-    HomogeneousMediumComponent() = default;
-
-    glm::vec3 albedo   = glm::vec3(0);
-    float density      = 0;
-    float g            = 0.0f;
-    float Le           = 0.0f;
-    glm::vec3 Le_color = glm::vec3(1);
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Homogeneous Medium"; }
-};
-
-struct HeterogeneousMediumComponent
-{
-    HeterogeneousMediumComponent()
-    {
-        // Create default 1x1x1 textures for density, albedo and emission
-        TextureSpecification spec;
-        spec.width  = 1;
-        spec.height = 1;
-        spec.depth  = 1;
-        spec.format = TextureFormat::RGBFLOAT;
-        glm::vec3 black(0);
-        glm::vec3 white(1.0f);
-        _default_emission_texture = atcg::Texture3D::create(&black, spec);
-
-        _default_albedo_texture = atcg::Texture3D::create(&white, spec);
-    }
-
-    ATCG_INLINE atcg::ref_ptr<Texture3D> density() const
-    {
-        return AssetManager::getAsset<Texture3D>(density_grid.handle);
-    }
-    ATCG_INLINE atcg::ref_ptr<Texture3D> albedo() const
-    {
-        return AssetManager::isAssetHandleValid(albedo_grid.handle)
-                   ? AssetManager::getAsset<Texture3D>(albedo_grid.handle)
-                   : _default_albedo_texture;
-    }
-    ATCG_INLINE atcg::ref_ptr<Texture3D> emission() const
-    {
-        return AssetManager::isAssetHandleValid(emission_grid.handle)
-                   ? AssetManager::getAsset<Texture3D>(emission_grid.handle)
-                   : _default_emission_texture;
-    }
-
-    struct GridComponent
-    {
-        BoundingBox bbox;
-        AssetHandle handle = 0;
-        float scale        = 1.0f;
-    };
-
-    GridComponent density_grid;
-    GridComponent albedo_grid;
-    GridComponent emission_grid;
-
-    float g = 0.0f;
-
-    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Heterogeneous Medium"; }
-
-private:
-    atcg::ref_ptr<Texture3D> _default_emission_texture;
-    atcg::ref_ptr<Texture3D> _default_albedo_texture;
-};
-
-}    // namespace atcg
+#include <Scene/Components/AccelerationStructureComponent.h>
+#include <Scene/Components/CameraComponent.h>
+#include <Scene/Components/CustomRenderComponent.h>
+#include <Scene/Components/EdgeCylinderRenderComponent.h>
+#include <Scene/Components/EdgeRenderComponent.h>
+#include <Scene/Components/EditorCameraComponent.h>
+#include <Scene/Components/GeometryComponent.h>
+#include <Scene/Components/HeterogeneousMediumComponent.h>
+#include <Scene/Components/HomogeneousMediumComponent.h>
+#include <Scene/Components/IDComponent.h>
+#include <Scene/Components/InstanceRenderComponent.h>
+#include <Scene/Components/MeshLightComponent.h>
+#include <Scene/Components/MeshRenderComponent.h>
+#include <Scene/Components/NameComponent.h>
+#include <Scene/Components/PointLightComponent.h>
+#include <Scene/Components/PointRenderComponent.h>
+#include <Scene/Components/PointSphereRenderComponent.h>
+#include <Scene/Components/RenderComponent.h>
+#include <Scene/Components/ScriptComponent.h>
+#include <Scene/Components/TransformComponent.h>
