+++ atcg_lib/include/Scene/Components/CustomRenderComponent.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <Renderer/Camera.h>
+#include <Scene/Components/RenderComponent.h>
+#include <Scene/Entity.h>
+
+namespace atcg
+{
+struct CustomRenderComponent : public RenderComponent
+{
+    using RenderCallbackFn = std::function<void(Entity, const atcg::ref_ptr<Camera>& camera)>;
+
+    CustomRenderComponent(const RenderCallbackFn& callback) : RenderComponent(), callback(callback) {}
+
+    RenderCallbackFn callback;
+};
+}    // namespace atcg
