+++ atcg_lib/include/Scene/Components/InstanceRenderComponent.h
@@ -1,0 +1,35 @@
+#pragma once
+
+#include <Asset/AssetManagerSystem.h>
+#include <Scene/Components/RenderComponent.h>
+
+namespace atcg
+{
+struct InstanceRenderComponent : public RenderComponent
+{
+    InstanceRenderComponent() : RenderComponent() {}
+
+    void addInstanceBuffer(const atcg::ref_ptr<VertexBuffer>& instance_vbo) { instance_vbos.push_back(instance_vbo); }
+
+    ATCG_INLINE atcg::ref_ptr<Material> material() const
+    {
+        auto mat = AssetManager::getAsset<Material>(material_handle);
+        return mat ? mat : default_material;
+    }
+
+    ATCG_INLINE atcg::ref_ptr<Shader> shader() const
+    {
+        auto shader = AssetManager::getAsset<Shader>(shader_handle);
+        return shader ? shader : default_shader;
+    }
+
+    static ATCG_CONSTEXPR ATCG_INLINE const char* toString() { return "Instance Renderer"; }
+
+    std::vector<atcg::ref_ptr<VertexBuffer>> instance_vbos;
+    atcg::ref_ptr<atcg::Shader> default_shader = nullptr;
+    bool receive_shadow                        = true;
+
+    AssetHandle material_handle = 0;
+    AssetHandle shader_handle;
+};
+}    // namespace atcg
