+++ atcg_lib/include/Scene/Components/NameComponent.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <string>
+
+namespace atcg
+{
+struct NameComponent
+{
+    NameComponent() = default;
+    NameComponent(const std::string& name) : _name(name) {}
+
+    ATCG_INLINE const std::string& name() const { return _name; }
+
+private:
+    std::string _name;
+};
+}    // namespace atcg
